updated data type for price

+++ DTOs/BookDTO.cs
@@ -20,7 +20,7 @@ namespace BookStore_API.DTOs
 
         public string Image { get; set; }
 
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
 
         public int? AuthorId { get; set; }
 
@@ -42,7 +42,7 @@ namespace BookStore_API.DTOs
 
         public string Image { get; set; }
 
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
 
         [Required]
         public int? AuthorId { get; set; }
@@ -63,7 +63,7 @@ namespace BookStore_API.DTOs
 
         public string Image { get; set; }
 
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
 
         public int? AuthorId { get; set; }
     }
